requirements.txt
psutil
requests
tenacity
mutagen
Pillow
pygame
nest-asyncio
tiktoklive